% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{analyzePlate}
\alias{analyzePlate}
\title{Analyze the results of a Perkin Elmer plate data}
\usage{
analyzePlate(filename, wells, samples, ref)
}
\arguments{
\item{filename}{String, name of the file with the data}

\item{wells}{Vector with well names}

\item{samples}{Vector with sample names}

\item{ref}{Reference sample}
}
\description{
Analyze the results of a Perkin Elmer plate data
}
\examples{
a = buildPlate("2106-05-08-15-51.txt",
               wells = rectWells("F05", "H12"),
               samples = c(1, 7, 4, NA, 3, 7, 6, 3,
                           2, 5, 1, 2, 7, 5, NA, 6,
                           4, 3, 5, 6, 2, 4, NA, 1),
               ref = 1)

}

