% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{checkWellNames}
\alias{checkWellNames}
\title{Check and convert well names to the appropriate format
E.g. A1 -> A01}
\usage{
checkWellNames(wellNames)
}
\arguments{
\item{wellNames}{String vector (e.g. c("A1", "A2", "B3", "B4"))}
}
\value{
String vector
}
\description{
Check and convert well names to the appropriate format
E.g. A1 -> A01
}

