% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{displayBioscreenWells}
\alias{displayBioscreenWells}
\title{Display some labels in the Bioscreen plate format}
\usage{
displayBioscreenWells(x, labels = NULL, wells = "well", col = NULL,
  lwd = NULL)
}
\arguments{
\item{x}{Data frame with wells and labels in columns}

\item{labels}{String, name of the column with the labels to display}

\item{wells}{String, name of the column with the well names.}

\item{col}{String, name of the column with the color values}
}
\description{
The format for the well names can be either purely numeric (from 1 to
100) or can used letter/number coordinates a la 96-well plate (from A01
to J10, letters for rows and numbers for columns).
}

