% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{displayWells}
\alias{displayWells}
\title{Display some labels in the 96-well plate format}
\usage{
displayWells(x, labels = NULL, wells = "well", col = NULL, lwd = NULL)
}
\arguments{
\item{x}{Data frame with wells and labels in columns}

\item{labels}{String, name of the column with the labels to display}

\item{wells}{String, name of the column with the well names}

\item{col}{String, name of the column with the color values}
}
\description{
Display some labels in the 96-well plate format
}

