% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{distribSamples}
\alias{distribSamples}
\title{Distribute replicated samples randomly across a 96-well plate}
\usage{
distribSamples(samples, replicates = 3, wells = NULL, blockLength = NULL,
  palette = "Set2", alpha.f = 0.5)
}
\arguments{
\item{samples}{Vector of string containing the sample names}

\item{replicates}{Integer, number of replicate to generate for each
sample name}

\item{wells}{Optional, vector containing the wells to be used, for
example generated using rectWells(). If NULL, uses the whole plate.
If blockLength is used, this has no effect.}

\item{blockLength}{Optional, can be either 2, 4 or 8. If given, tries 
to optimize distribution for use with multichannel pipette using
sample blocks of length blockLength}
}
\value{
Data frame with the plate plan
}
\description{
Distribute replicated samples randomly across a 96-well plate
}
\examples{
samples = c("c0", "c1", "c2", "c3", "c4", "c5", "c6", "w", "DNase+", "DNase-",
            "EGTA-", "incub90", "incub75", "incub60", "dil2", "dil4", "dil8",
            "dil16", "dil32", "toto1", "toto2", "toto3", "toto4", "a1", "a2",
            "a3", "a4", "a5", "b1", "b2", "b3", "b4")
d = distribSamples(samples, replicates = 3, blockLength = 4, palette = "Pastel1",
                   alpha.f = 1)
savePlan(d, "toto.pdf")

}

