% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{loadPerkinElmerFile}
\alias{loadPerkinElmerFile}
\title{Load the data from a Perkin Elmer text file. The text file will be
loaded until the first line with "End time"}
\usage{
loadPerkinElmerFile(filename)
}
\description{
Load the data from a Perkin Elmer text file. The text file will be
loaded until the first line with "End time"
}

