% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{mergeRepeats}
\alias{mergeRepeats}
\title{Merge the measurement repeats from a Perkin Elmer table}
\usage{
mergeRepeats(data)
}
\arguments{
\item{data}{Output from loadPerkinElmerFile}
}
\value{
List
}
\description{
Merge the measurement repeats from a Perkin Elmer table
}
\examples{
a = loadPerkinElmerFile("2106-05-08-15-51-data.txt")
b = mergeRepeats(a)

}

