% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{plateCoord}
\alias{plateCoord}
\title{Add a colPos and rowCol columns to a data frame with a well column}
\usage{
plateCoord(data)
}
\description{
Add a colPos and rowCol columns to a data frame with a well column
}

