% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{processPerkinElmerTable}
\alias{processPerkinElmerTable}
\title{Convert the raw data frame from Perkin Elmer to light table}
\usage{
processPerkinElmerTable(data)
}
\description{
Convert the raw data frame from Perkin Elmer to light table
}

