% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/96-wells-functions.R
\name{rectWells}
\alias{rectWells}
\title{Generate a vector with well names within a rectangular region of a plate}
\usage{
rectWells(topleft, bottomright, transpose = F)
}
\description{
Generate a vector with well names within a rectangular region of a plate
}
\examples{
rectWells("A01", "C06")

}

